Render errors as JSON in the API
A side note: I would like to eventually implement DRYer error handling
as shown in this article: http://blog.golang.org/error-handling-and-go.

Unfortunately, I'm not sure that strategy will work with mux, the HTTP
router we are using. I think the answer is Middleware, but whatever the
case, we need to be able to define our CRUD actions as funcs that return
an error. Then, in only one spot, we will control the formatting and
rendering of the error.