Restore cursor position in UI textarea input (#126)
Prior to this, the textarea would have its input pretty-printed and
replaced on keyup, which would cause the cursor to snap to the bottom...
which was particularly annoying. This behavior has been replaced with a
simple autotab-on-newline behavior. IMO this achieves the same thing --
enforcing pretty JSON. It just doesn't lose your cursor position.