Deploy to VPC (#195)
- This change allows a user to deploy a kubernetes cluster to an existing VPC.
This is done by specifying the "vpc_id" option with the VPC ID as a string.
Supergiant will then mark the VPC as not managed and deploy into the VPC. When
The cluster is deleted, it will skip destruction of the VPC. NOTE: User must be
aware to configure IP ranges correctly in the kube config during deploy. Missmatched
IP ranges will result in a build error.