Small update to AWS subnet public DNS during VPC injection.
If your kubernetes cluster is accessable by the outside, you must manually enable
public DNS before build time.

If it is not accessable, you must be sure to have it disabled.

SG will get confused and not know when IP to hit during operation if this
is not manaully set.