Helm Chart Form Update.
This change updates a bug in helm charts, that was not
updateing charts that have new versions. There are still
a large number of charts in the kubernetes "stable" helm
repo that simply do not work without modification. :-/
The solution may be to not make the kubernetes repository
the default repo, in favor of an actually stable repo...